refactor : one-to-one 컴포넌트 도메인 폴더 통합
archive, balance-game, hall-of-fame, study-history, interview,
one-on-one, calendars 폴더를 src/components/one-to-one/ 하위로 통합.
관련 import 경로 일괄 수정.
