+++ src/app/(service)/one-on-one/page.tsx
@@ -1,4 +1,4 @@
-import OneOnOnePage from '@/components/one-on-one/one-on-one-page';
+import OneOnOnePage from '@/components/one-to-one/sessions/one-on-one-page';
 
 export default function OneOnOnePageRoute() {
   return <OneOnOnePage />;
