+++ src/components/one-to-one/archive/archive-pagination.tsx
@@ -1,19 +1,20 @@
 'use client';
 
+import { ChevronLeft, ChevronRight } from 'lucide-react';
 import React from 'react';
-import PaginationCircleButton from '@/components/one-on-one/pagination-circle-button';
+import PaginationCircleButton from '@/components/one-to-one/sessions/pagination-circle-button';
 
-interface StudyHistoryPaginationProps {
+interface ArchivePaginationProps {
   currentPage: number;
   totalPages: number;
   onPageChange: (page: number) => void;
 }
 
-export default function StudyHistoryPagination({
+export default function ArchivePagination({
   currentPage,
   totalPages,
   onPageChange,
-}: StudyHistoryPaginationProps) {
+}: ArchivePaginationProps) {
   if (totalPages <= 1) return null;
 
   return (
@@ -22,7 +23,7 @@ export default function StudyHistoryPagination({
         onClick={() => onPageChange(Math.max(1, currentPage - 1))}
         disabled={currentPage === 1}
       >
-        ←
+        <ChevronLeft className="h-5 w-5" />
       </PaginationCircleButton>
       <span className="font-designer-15m text-text-subtle bg-background-default border-border-subtle flex h-[40px] items-center justify-center rounded-[9999px] border px-300">
         {currentPage} / {totalPages}
@@ -31,7 +32,7 @@ export default function StudyHistoryPagination({
         onClick={() => onPageChange(Math.min(totalPages, currentPage + 1))}
         disabled={currentPage === totalPages}
       >
-        →
+        <ChevronRight className="h-5 w-5" />
       </PaginationCircleButton>
     </div>
   );
