+++ src/components/one-to-one/balance-game/community-tab-client.tsx
@@ -3,9 +3,9 @@
 import { Loader2, Vote, SearchX, Plus, MessageSquareText } from 'lucide-react';
 import { useRouter, useSearchParams } from 'next/navigation';
 import React, { useState, useEffect, useMemo } from 'react';
-import VotingCreateModal from '@/components/balance-game/voting/voting-create-modal';
-import VotingDetailView from '@/components/balance-game/voting/voting-detail-view';
-import VotingCard from '@/components/balance-game/voting-card';
+import VotingCreateModal from '@/components/one-to-one/balance-game/voting/voting-create-modal';
+import VotingDetailView from '@/components/one-to-one/balance-game/voting/voting-detail-view';
+import VotingCard from '@/components/one-to-one/balance-game/voting-card';
 import SectionHeader from '@/components/common/ui/section-header';
 import SectionShell from '@/components/common/ui/section-shell';
 import Toast from '@/components/common/ui/toast';
