+++ src/components/one-to-one/balance-game/voting-card.tsx
