+++ src/components/one-to-one/balance-game/voting/daily-stats-chart.tsx
