+++ src/components/one-to-one/balance-game/voting/vote-results-chart.tsx
