+++ src/components/one-to-one/balance-game/voting/vote-timer.tsx
