+++ src/components/one-to-one/balance-game/voting/voting-detail-modal.tsx
