+++ src/components/one-to-one/balance-game/voting/voting-modal-footer.tsx
