+++ src/components/one-to-one/balance-game/voting/voting-modal-header.tsx
