+++ src/components/one-to-one/balance-game/voting/voting-option-fields.tsx
