+++ src/components/one-to-one/balance-game/voting/voting-tag-field.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React from 'react';
-import TagAutocomplete from '@/components/balance-game/tag-autocomplete';
+import TagAutocomplete from '@/components/one-to-one/balance-game/tag-autocomplete';
 import {
   BALANCE_GAME_TAG_MAX_COUNT,
   BALANCE_GAME_TAG_MAX_LEN,
