+++ src/components/one-to-one/hall-of-fame/mvp-team-card.tsx
