+++ src/components/one-to-one/history/study-history-pagination.tsx
@@ -1,20 +1,19 @@
 'use client';
 
-import { ChevronLeft, ChevronRight } from 'lucide-react';
 import React from 'react';
-import PaginationCircleButton from '@/components/one-on-one/pagination-circle-button';
+import PaginationCircleButton from '@/components/one-to-one/sessions/pagination-circle-button';
 
-interface ArchivePaginationProps {
+interface StudyHistoryPaginationProps {
   currentPage: number;
   totalPages: number;
   onPageChange: (page: number) => void;
 }
 
-export default function ArchivePagination({
+export default function StudyHistoryPagination({
   currentPage,
   totalPages,
   onPageChange,
-}: ArchivePaginationProps) {
+}: StudyHistoryPaginationProps) {
   if (totalPages <= 1) return null;
 
   return (
@@ -23,7 +22,7 @@ export default function ArchivePagination({
         onClick={() => onPageChange(Math.max(1, currentPage - 1))}
         disabled={currentPage === 1}
       >
-        <ChevronLeft className="h-5 w-5" />
+        ←
       </PaginationCircleButton>
       <span className="font-designer-15m text-text-subtle bg-background-default border-border-subtle flex h-[40px] items-center justify-center rounded-[9999px] border px-300">
         {currentPage} / {totalPages}
@@ -32,7 +31,7 @@ export default function ArchivePagination({
         onClick={() => onPageChange(Math.min(totalPages, currentPage + 1))}
         disabled={currentPage === totalPages}
       >
-        <ChevronRight className="h-5 w-5" />
+        →
       </PaginationCircleButton>
     </div>
   );
