+++ src/components/pages/home-page-server-content.tsx
@@ -1,9 +1,9 @@
 import { Suspense } from 'react';
-import ArchiveTab from '@/components/archive/archive-tab';
-import CommunityTab from '@/components/balance-game/community-tab';
-import HallOfFameTab from '@/components/hall-of-fame/hall-of-fame-tab';
+import ArchiveTab from '@/components/one-to-one/archive/archive-tab';
+import CommunityTab from '@/components/one-to-one/balance-game/community-tab';
+import HallOfFameTab from '@/components/one-to-one/hall-of-fame/hall-of-fame-tab';
 import StudyTab from '@/components/group-study/schedule/home-study-tab';
-import StudyHistoryTab from '@/components/study-history/study-history-tab';
+import StudyHistoryTab from '@/components/one-to-one/history/study-history-tab';
 import { isAuthenticatedMemberSessionState } from '@/features/auth/model/auth-session';
 import { readServerAuthSession } from '@/features/auth/model/server-auth-session';
 import type { HomeTab } from '@/features/home/model/home-page-search-params';
